requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
psycopg2-binary==2.9.9
pyodbc==5.0.1
cx_Oracle==8.3.0
python-dotenv==1.0.1
streamlit==1.31.1
requests==2.31.0
pydantic==2.6.1
pydantic[email]==2.6.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2 
mistralai==1.7.0